+++ test/reporting.test.jl
@@ -1,81 +1,81 @@
-module Reporting
-
-using Jute
 using TestUtils
 
 
-TESTCASES = Dict(
-    :multiple_tests => testcase() do
+@testgroup "reporting" begin
+
+
+TESTCASES = [
+    testcase("multiple tests") do
         @test 1 == 1
         @test 2 == 2
         @test 3 == 3
     end,
 
-    :returning_value => testcase() do
+    testcase("returning value") do
         @test 1 == 1
         @test_result 10
     end,
 
-    :multiple_tests_and_one_failure => testcase() do
+    testcase("multiple tests and one failure") do
         @test 1 == 1
         @test 2 == 1
         @test 3 == 3
     end,
 
-    :uncaught_exception => testcase() do
+    testcase("uncaught exception") do
         @test 1 == 1
         error("Uncaught exception")
         @test 1 == 1
     end,
 
-    :caught_exception => testcase() do
+    testcase("caught exception") do
         @test 1 == 1
         @test_throws ErrorException error("Caught exception")
         @test 1 == 1
     end,
 
-    :skip_test => testcase() do
+    testcase("skip test") do
         @test 1 == 1
         @test_skip 1 == 2
         @test 1 == 1
     end,
 
-    :expected_failure => testcase() do
+    testcase("expected failure") do
         @test 1 == 1
         @test_broken 1 == 2
         @test 1 == 1
     end,
 
-    :unexpected_pass => testcase() do
+    testcase("unexpected pass") do
         @test 1 == 1
         @test_broken 1 == 1
         @test 1 == 1
     end,
 
-    :with_fixtures => testcase([1], [2]) do x, y
-    end
-    )
+    testcase("with fixtures", [1], [2]) do x, y
+    end,
+    ]
 
 
-verbosity0 = testcase() do
+@testcase "verbosity0" begin
     exitcode, output = nested_run_with_output(TESTCASES, Dict(:verbosity => 0))
     @test exitcode == 1
 
     template = """
         ================================================================================
-        multiple_tests_and_one_failure
+        multiple tests and one failure
         Test Failed
           Expression: 2 == 1
            Evaluated: 2 == 1
         ================================================================================
-        uncaught_exception
+        uncaught exception
         Error During Test
           Got an exception of type ErrorException outside of a @test
           Uncaught exception
           Stacktrace:
         <<<MULTILINE>>>
         ================================================================================
-        unexpected_pass
+        unexpected pass
         Error During Test
          Unexpected Pass
          Expression: 1 == 1
@@ -86,7 +86,7 @@ verbosity0 = testcase() do
 end
 
 
-verbosity1 = testcase() do
+@testcase "verbosity1" begin
     exitcode, output = nested_run_with_output(TESTCASES, Dict(:verbosity => 1))
     @test exitcode == 1
 
@@ -100,19 +100,19 @@ verbosity1 = testcase() do
         --------------------------------------------------------------------------------
         18 tests passed, 1 failed, 2 errored in <<<full_time>>> (total test time <<<test_time>>>)
         ================================================================================
-        multiple_tests_and_one_failure
+        multiple tests and one failure
         Test Failed
           Expression: 2 == 1
            Evaluated: 2 == 1
         ================================================================================
-        uncaught_exception
+        uncaught exception
         Error During Test
           Got an exception of type ErrorException outside of a @test
           Uncaught exception
           Stacktrace:
         <<<MULTILINE>>>
         ================================================================================
-        unexpected_pass
+        unexpected pass
         Error During Test
          Unexpected Pass
          Expression: 1 == 1
@@ -123,7 +123,7 @@ verbosity1 = testcase() do
 end
 
 
-verbosity2 = testcase() do
+@testcase "verbosity2" begin
     exitcode, output = nested_run_with_output(TESTCASES, Dict(:verbosity => 2))
     @test exitcode == 1
 
@@ -133,31 +133,31 @@ verbosity2 = testcase() do
         ================================================================================
         Platform: Julia <<<julia_version>>>, Jute <<<jute_version>>>
         --------------------------------------------------------------------------------
-        multiple_tests (<<<time>>>) [PASS] [PASS] [PASS]
-        returning_value (<<<time>>>) [PASS] [10]
-        multiple_tests_and_one_failure (<<<time>>>) [PASS] [FAIL] [PASS]
-        uncaught_exception (<<<time>>>) [PASS] [ERROR]
-        caught_exception (<<<time>>>) [PASS] [PASS] [PASS]
-        skip_test (<<<time>>>) [PASS] [BROKEN] [PASS]
-        expected_failure (<<<time>>>) [PASS] [BROKEN] [PASS]
-        unexpected_pass (<<<time>>>) [PASS] [ERROR] [PASS]
-        with_fixtures[1,2] (<<<time>>>) [PASS]
+        multiple tests (<<<time>>>) [PASS] [PASS] [PASS]
+        returning value (<<<time>>>) [PASS] [10]
+        multiple tests and one failure (<<<time>>>) [PASS] [FAIL] [PASS]
+        uncaught exception (<<<time>>>) [PASS] [ERROR]
+        caught exception (<<<time>>>) [PASS] [PASS] [PASS]
+        skip test (<<<time>>>) [PASS] [BROKEN] [PASS]
+        expected failure (<<<time>>>) [PASS] [BROKEN] [PASS]
+        unexpected pass (<<<time>>>) [PASS] [ERROR] [PASS]
+        with fixtures[1,2] (<<<time>>>) [PASS]
         --------------------------------------------------------------------------------
         18 tests passed, 1 failed, 2 errored in <<<full_time>>> (total test time <<<test_time>>>)
         ================================================================================
-        multiple_tests_and_one_failure
+        multiple tests and one failure
         Test Failed
           Expression: 2 == 1
            Evaluated: 2 == 1
         ================================================================================
-        uncaught_exception
+        uncaught exception
         Error During Test
           Got an exception of type ErrorException outside of a @test
           Uncaught exception
           Stacktrace:
         <<<MULTILINE>>>
         ================================================================================
-        unexpected_pass
+        unexpected pass
         Error During Test
          Unexpected Pass
          Expression: 1 == 1
@@ -168,26 +168,26 @@ verbosity2 = testcase() do
 end
 
 
-captured_output = testcase() do
-    testcases = Dict(
-        :passing_tc => testcase() do
+@testcase "captured_output" begin
+    testcases = [
+        testcase("passing testcase") do
             println(STDOUT, "stdout from passing testcase")
             println(STDERR, "stderr from passing testcase")
         end,
 
-        :failing_tc => testcase() do
+        testcase("failing testcase") do
             println(STDOUT, "stdout from failing testcase")
             @test 1 == 2
             println(STDERR, "stderr from failing testcase")
         end,
-        )
+        ]
 
     exitcode, output = nested_run_with_output(
         testcases, Dict(:verbosity => 0, :capture_output => true))
 
     template = """
         ================================================================================
-        failing_tc
+        failing testcase
         Captured output:
         stdout from failing testcase
         stderr from failing testcase